requirements.txt
python-telegram-bot[job-queue]==21.5
yt-dlp>=2024.12.23
ytmusicapi==1.7.4
shazamio==0.6.0
aiohttp==3.9.5
